+++ plugins/org.jkiss.dbeaver.ext.turbographpp.graph/META-INF/MANIFEST.MF
@@ -24,4 +24,5 @@ Bundle-ClassPath: .,
  lib/JavaFXSmartGraph-0.9.4.jar
 Automatic-Module-Name: org.jkiss.dbeaver.ext.turbographpp.graph
 Export-Package: org.jkiss.dbeaver.ext.turbographpp.graph,
+ org.jkiss.dbeaver.ext.turbographpp.graph.data,
  org.jkiss.dbeaver.ext.turbographpp.graph.test
