+++ plugins/org.jkiss.dbeaver.ext.turbographpp.graph/src/org/jkiss/dbeaver/ext/turbographpp/graph/internal/GraphResources.properties
@@ -1,0 +1,9 @@
+fxgraph_export_csv_dialog_title = Export CSV
+fxgraph_export_csv_dialog_default_msg = Please Select Folder
+fxgraph_export_csv_dialog_error_msg_select_folder = Error, Please Select Folder
+fxgraph_export_csv_dialog_error_msg_inpt_filename = Error, Please input FileName
+fxgraph_export_csv_dialog_label_folder = Select Folder
+fxgraph_export_csv_dialog_label_filename = File Name
+fxgraph_export_csv_dialog_label_node = Node Files
+fxgraph_export_csv_dialog_label_edge = Edge Files
+fxgraph_export_csv_dialog_directory_dialog_title = Please select directory to save.
