+++ plugins/org.jkiss.dbeaver.ext.turbographpp.graph/src/org/jkiss/dbeaver/ext/turbographpp/graph/internal/GraphResources_kr.properties
@@ -1,0 +1,9 @@
+fxgraph_export_csv_dialog_title = CSV 내보내기
+fxgraph_export_csv_dialog_default_msg = 저장할 폴더를 선택해 주세요.
+fxgraph_export_csv_dialog_error_msg_select_folder = 폴더는 선택하세요
+fxgraph_export_csv_dialog_error_msg_inpt_filename = 파일 이름을 입력하세요.
+fxgraph_export_csv_dialog_label_folder = 폴더
+fxgraph_export_csv_dialog_label_filename = 파일이름
+fxgraph_export_csv_dialog_label_node = 정점 파일들
+fxgraph_export_csv_dialog_label_edge = 간선 파일들
+fxgraph_export_csv_dialog_directory_dialog_title = 저장할 폴더를 선택하세요.
